requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==1.0.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
